Move replication variables test to dedicated module
