+++ tests/spread/integration/test_replication_variables.py/task.yaml
@@ -1,0 +1,7 @@
+summary: test_replication_variables.py
+environment:
+  TEST_MODULE: high_availability/test_replication_variables.py
+execute: |
+  tox run -e integration -- "tests/integration/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
+artifacts:
+  - allure-results
